# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: SalesHub sales API
  description: 'Esta API permite gestionar tipos de pago,historial de ventas y generacion de ventas'
  version: 1.0.0
servers:
  - url: 'http://localhost:8013/servicesales/v1'
x-devops-properties:
  apiType: ne
  apiDomain: custom
  apiName: servicesales
  apiPlatform: oc
tags:
  - name: servicesales-controller
    description: Service sales Management Controller
paths:
  /sales:
    post:
      tags:
        - servicesales-controller
      summary: Create a new sale in the service sales management system(createSales)
      description: Generacion de una nueva venta
      operationId:  createSales
      parameters:
        - $ref: "#/components/parameters/unicaServiceId"
        - $ref: "#/components/parameters/unicaApplication"
        - $ref: "#/components/parameters/unicaPid"
        - $ref: "#/components/parameters/unicaUser"
        - $ref: "#/components/parameters/authorization"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleRequest'
      responses:
        '201':
          description: Created
          headers:
            UNICA-ServiceId:
              $ref: '#/components/headers/UNICA-ServiceId'
            UNICA-PID:
              $ref: '#/components/headers/UNICA-PID'
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sales/history:
    get:
      tags:
        - servicesales-controller
      summary: Consult the history of sales(retrieveHistorySales)
      description: Historial de Ventas
      operationId: retrieveHistorySales
      parameters:
        - $ref: "#/components/parameters/unicaServiceId"
        - $ref: "#/components/parameters/unicaApplication"
        - $ref: "#/components/parameters/unicaPid"
        - $ref: "#/components/parameters/unicaUser"
        - $ref: "#/components/parameters/authorization"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveSalesHistoryResponse'
          headers:
            UNICA-ServiceId:
              $ref: "#/components/headers/UNICA-ServiceId"
            UNICA-ServicePID:
              $ref: "#/components/headers/UNICA-PID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sales/history/detail:
    get:
      tags:
        - servicesales-controller
      summary: Consult the list of sales detail(retrieveSalesDetail)
      description: Lista de detalle de venta
      operationId: retrieveSalesDetail
      parameters:
        - $ref: "#/components/parameters/unicaServiceId"
        - $ref: "#/components/parameters/unicaApplication"
        - $ref: "#/components/parameters/unicaPid"
        - $ref: "#/components/parameters/unicaUser"
        - $ref: "#/components/parameters/authorization"
        - $ref: "#/components/parameters/filter"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveSalesDetailResponse'
          headers:
            UNICA-ServiceId:
              $ref: "#/components/headers/UNICA-ServiceId"
            UNICA-ServicePID:
              $ref: "#/components/headers/UNICA-PID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sales/payment-type:
    get:
      tags:
        - servicesales-controller
      summary: Consult the list of paymentType(retrievePaymentType)
      description: Lista de tipo de pago en una venta
      operationId: retrievePaymentType
      parameters:
        - $ref: "#/components/parameters/unicaServiceId"
        - $ref: "#/components/parameters/unicaApplication"
        - $ref: "#/components/parameters/unicaPid"
        - $ref: "#/components/parameters/unicaUser"
        - $ref: "#/components/parameters/authorization"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievePaymentType'
          headers:
            UNICA-ServiceId:
              $ref: "#/components/headers/UNICA-ServiceId"
            UNICA-ServicePID:
              $ref: "#/components/headers/UNICA-PID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - servicesales-controller
      summary: Create a new paymentType in the service sales management system(createPaymentType)
      description: Generacion de un nuevo tipo de pago
      operationId:  createPaymentType
      parameters:
        - $ref: "#/components/parameters/unicaServiceId"
        - $ref: "#/components/parameters/unicaApplication"
        - $ref: "#/components/parameters/unicaPid"
        - $ref: "#/components/parameters/unicaUser"
        - $ref: "#/components/parameters/authorization"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Created
          headers:
            UNICA-ServiceId:
              $ref: '#/components/headers/UNICA-ServiceId'
            UNICA-PID:
              $ref: '#/components/headers/UNICA-PID'
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /sales/payment-type/{paymentTypeId}:
    patch:
      tags:
        - servicesales-controller
      summary: Update payment by id
      description: Actualiza el tipo de pago por su identificador
      operationId: updatePaymentType
      parameters:
        - $ref: "#/components/parameters/unicaServiceId"
        - $ref: "#/components/parameters/unicaApplication"
        - $ref: "#/components/parameters/unicaPid"
        - $ref: "#/components/parameters/unicaUser"
        - $ref: "#/components/parameters/authorization"
        - name: paymentTypeId
          in: path
          required: true
          schema:
            type: integer
            description: paymentType id
            example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaymentTypeRequest'
      responses:
        '204':
          description: No Content
          headers:
            UNICA-ServiceId:
              $ref: "#/components/headers/UNICA-ServiceId"
            UNICA-PID:
              $ref: "#/components/headers/UNICA-PID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  schemas:
    CreateSaleRequest:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductRequest'
        paymentTypeId:
          type: integer
          description: Id of payment type
          example: 5
        aditionalData:
          type: array
          items:
            $ref: '#/components/schemas/AditionalData'
    ProductRequest:
      type: object
      properties:
        productId:
          type: integer
          description: Id of product
          example: 1
        amount:
          type: integer
          description: Amount of product
          example: 5
    RetrieveSalesHistoryResponse:
      type: object
      description: Retrieve Sales history Response
      properties:
        sales:
          type: array
          items:
            $ref: '#/components/schemas/SalesResponse'
        metadata:
          $ref: '#/components/schemas/MetadataResponse'
    SalesResponse:
      type: object
      properties:
        id:
          type: integer
          description: Id of sale
          example: 1
        registerDate:
          type: string
          description: register date of sale
          example: "15/03/2025"
        paymentTypeDescription:
          type: string
          description: type of payment
          example: "PAYPAL"
        total:
          type: number
          description: total of sale
          example: 159.99
        product:
          type: string
          description: "Description of the product."
          example: "description product"
        amount:
          type: integer
          description: "Amount of product"
          example: 1
        priceUnit:
          type: number
          description: "Price Unit"
          example: 19.99
    RetrieveSalesDetailResponse:
      type: object
      description: Retrieve SalesDetail Response
      properties:
        salesDetail:
          type: array
          items:
            $ref: '#/components/schemas/SaleDetailResponse'
    SaleDetailResponse:
      type: object
      properties:
        product:
          type: string
          description: "Description of the product."
          example: "description product"
        amount:
          type: integer
          description: "Amount of product"
          example: 1
        priceUnit:
          type: number
          description: "Price Unit"
          example: 19.99
        total:
          type: number
          description: "Total"
          example: 38.99
        additionalData:
          type: array
          items:
            $ref: '#/components/schemas/AditionalData'
    RetrievePaymentType:
      type: object
      description: Retrieve SalesDetail Response
      properties:
        paymentsType:
          type: array
          items:
            $ref: '#/components/schemas/PaymentTypeResponse'
        metadata:
          $ref: '#/components/schemas/MetadataResponse'
    PaymentTypeResponse:
      type: object
      properties:
        id:
          type: integer
          description: "Payment Type Id"
          example: 1
        description:
          type: string
          description: "Payment Type Description"
          example: "paypal"
        status:
          type: integer
          description: "Payment Type status"
          example: 1
        additionalData:
          type: array
          items:
            $ref: '#/components/schemas/AditionalData'
    CreatePaymentRequest:
      type: object
      properties:
        description:
          type: string
          description: "Payment Type Description"
          example: "paypal"
        status:
          type: integer
          description: "Payment Type status"
          example: 1
        additionalData:
          type: array
          items:
            $ref: '#/components/schemas/AditionalData'
    UpdatePaymentTypeRequest:
      type: object
      properties:
        description:
          type: string
          description: "Payment Type Description"
          example: "paypal"
        status:
          type: integer
          description: "Payment Type status"
          example: 1
        additionalData:
          type: array
          items:
            $ref: '#/components/schemas/AditionalData'
    MetadataResponse:
      type: object
      properties:
        limit:
          description: Limit
          type : integer
          format: int64
          example: 25
        offset:
          description: Offset
          type : integer
          format: int64
          example: 0
        totalElements:
          description: Total elements
          type : integer
          format: int64
          example: 10
    AditionalData:
      type: object
      description: Adittional data
      properties:
        key:
          type: string
          example: "aditional Key"
          description: "Clave de datos adicionales. Referencia (Key)"
        value:
          type: string
          example: "aditioanal value"
          description: "Valor de datos adicionales. Referencia (Value)"
    ErrorType:
      type: object
      properties:
        exceptionId:
          type: string
        exceptionText:
          type: string
        moreInfo:
          type: string
        userMessage:
          type: string
        httpStatus:
          type: integer
        exceptionDetails:
          type: array
          items:
            $ref: '#/components/schemas/ErrorTypeDetail'
      required:
        - exceptionId
        - exceptionText
        - moreInfo
        - userMessage
    ErrorTypeDetail:
      type: object
      properties:
        code:
          type: string
        component:
          type: string
        description:
          type: string
  parameters:
    unicaServiceId:
      name: UNICA-ServiceId
      in: header
      description: "Identificador único de cada ejecución dentro del contexto de la aplicación consumidora. Patrón: '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'."
      required: true
      schema:
        type: string
      example: 1b567df3-0fa8-4ad2-ab0b-a97291904361
    unicaApplication:
      name: UNICA-Application
      in: header
      description: "Identificación del sistema que origina la petición. "
      required: true
      schema:
        type: string
      example: wappe
    unicaPid:
      name: UNICA-PID
      in: header
      description: "Identificador único del flujo de ejecución para cubrir funcionalmente las interacciones que se puedan necesitar para completar un proceso o subproceso de negocio. Patrón: '[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}'. "
      required: true
      schema:
        type: string
      example: 550e8400-e29b-41d4-a716-446655440000
    unicaUser:
      name: UNICA-User
      in: header
      description: "Identificador del usuario del sistema y/o subsistema que inicia la petición. "
      required: true
      schema:
        type: string
      example: admin
    authorization:
      name: Authorization
      in: header
      description: "Incluye la prueba de acceso (usando el modelo de seguridad OAuth2.0) para garantizar que el consumidor tiene privilegios para ingresar a la base de datos."
      required: false
      schema:
        type: string
      example: 2q5jkdf3-0das8-4gd3-a12d-at3h5g357r46
    filter:
      name: filter
      in: query
      description: Filtros para la consulta
      required: false
      schema:
        type: string
        example: "code1:value1,code2:value2"
    sort:
      name: sort
      in: query
      description: Ordenar los resultados por el campo `nombre` en orden descendente.
      required: false
      schema:
        type: string
        example: "nombre,desc"
    limit:
      name: limit
      in: query
      description: Número máximo de productos a devolver.
      required: false
      schema:
        type: integer
        example: 15
    offset:
      name: offset
      in: query
      description: Desplazamiento de la página (para paginación)
      required: false
      schema:
        type: integer
        example: 0
  responses:
    BadRequest:
      description: "Bad Request"
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorType"
            example:
              exceptionId: "SVC0001"
              exceptionText: "Generic Client Error: client error"
              moreInfo: "API Generic wildcard fault response"
              userMessage: "Generic Client Error"
      headers:
        UNICA-ServiceId:
          $ref: "#/components/headers/UNICA-ServiceId"
        UNICA-PID:
          $ref: "#/components/headers/UNICA-PID"
    NotFound:
      description: "Not Found"
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorType"
            example:
              exceptionId: "SVC0004"
              exceptionText: "Requested API port or resource does not exist."
              moreInfo: "Requested API port or resource does not exist."
              userMessage: "Invalid Request URI"
      headers:
        UNICA-ServiceId:
          $ref: "#/components/headers/UNICA-ServiceId"
        UNICA-PID:
          $ref: "#/components/headers/UNICA-PID"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorType"
            example:
              exceptionId: "SVC0005"
              exceptionText: “Correlator %1 specified infield %2 is a duplicate”, %1Correlator, %2 - Field
              moreInfo: "Correlator specified in a request message is invalid because it already exists."
              userMessage: "Duplicated correlator"
      headers:
        UNICA-ServiceId:
          $ref: "#/components/headers/UNICA-ServiceId"
        UNICA-PID:
          $ref: "#/components/headers/UNICA-PID"
    InternalServerError:
      description: "Internal Server Error"
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorType"
            example:
              exceptionId: "SVR1000"
              exceptionText: "'Generic Server Error: server error"
              moreInfo: "There was a problem in the Service Providers network that prevented to carry out the request."
              userMessage: "Generic Server Fault"
      headers:
        UNICA-ServiceId:
          $ref: "#/components/headers/UNICA-ServiceId"
        UNICA-PID:
          $ref: "#/components/headers/UNICA-PID"
  headers:
    UNICA-ServiceId:
      description: "If this API is used via a platform acting as a common entry point to different OBs, this identifier is used to route the request to the corresponding OB environment"
      schema:
        type: string
    UNICA-PID:
      description: Unique identifier for the process or execution flow
      schema:
        type: string